Correct instantiation of AWS session object
The region needs to be set on the config object before creating the
session so that things like setting `AWS_STS_REGIONAL_ENDPOINTS` can
take effect.

Also, if credentials cannot be acquired, die at that point instead of
carrying on.
